workshop/Workplace: enable non-blocking mode on pipe and socket
LogBridge::Flush() / PipeLineReader::Flush() read from the pipe even
though it was not explicitly indicated as readable.

It can happen that the main process of a job exits, and
OnChildProcessExit() calls Flush(), which reads from the pipe.  If
another process of the job just happens to be still running, reading
from the pipe can block under rare circumstances, effectively causing
the Workshop daemon process to hang.

To avoid this, all file/socket descriptors handled inside the main
Workshop process should be non-blocking.