workshop/ProgressReader: fix double close bug
Fixes regression by commit 43afec463150590ce530bc7041479a3dfd5a8f8f
which caused hanging Workshop processes due to file descriptor
corruption after it was closed twice.

That commit caused the file descriptor to be copied (not moved) from
the SocketEvent to a local variable, which was then closed; however
the file descriptor was still registered in the SocketEvent, and will
be closed again in the destructor.  If somebody else just happens to
occupy that file descriptor number meanwhile, it will be closed,
and can be allocated again by somebody else, causing all kinds of
strange behavior.

The simple fix is to close the SocketEvent instead of the
FileDescriptor copy.  This way, the SocketEvent can properly update
its internal state, and makes the second SocketEvent::Close() call in
the destructor a no-op.