io/StateDirectories: new library
This allows enabling or disabling the cron queues by writing to 1 or 0
to "/run/cm4all/state/workshop/NAME/enabled" and sending SIGHUP.  We
will use this to mark inactive replica clusters as active or inactive.

So far, this feature is experimental.  Once the feature is stabilized,
other daemons (and documentation and command-line tools) will follow.