drop inheritable capabilities
When we switched to capabilities in commit df8b543995f, we instructed
systemd to give us the capabilities we needed; but these were now in
the "inheritable" set, which would also be inherited to spawned
processes.  Suddenly, all jobs had quais-superuser privileges, which
is a big vulnerability, of course.  Later, the spawner was modified to
drop inheritable capabilities (commit
https://github.com/CM4all/libcommon/commit/fd39963cb05892), but the
legacy Workshop version 5 was never updated to this version of the
spawner; Workshop 6.x is safe.